refactor: move collapsed paths state into TreeView and fix viewModes persistence
